+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 1.0.1
+* Add CircleCI
+* Refactored buildscript to do both `build` and `publish`
+* Add `.dockerignore`, and a check in the buildscript to ensure it exists
+
 ## 1.0.0
 * Init commit
 * Add API container image